feat: 중복 저장 안되도록 변경

+++ cat-jjal-maker-cra/src/App.js
@@ -44,9 +44,11 @@ const App = () => {
     }
 
     function handleHeartClick() {
-        const nextFavorites = [...favorites, mainCat];
-        setFavorites(nextFavorites);
-        jsonLocalStorage.setItem("favorites", nextFavorites);
+        if (!alreadyFavorite) {
+            const nextFavorites = [...favorites, mainCat];
+            setFavorites(nextFavorites);
+            jsonLocalStorage.setItem("favorites", nextFavorites);
+        }
     }
 
     const counterTitle = counter === null ? "" : counter + "번째 ";
